backend/compliance/requirements.txt
# Django Core 
Django==4.2.7 
djangorestframework==3.14.0 
 
# Database 
psycopg2-binary==2.9.9 
 
# CORS 
django-cors-headers==4.3.0 
 
# File Processing 
Pillow==10.1.0 
 
# Utilities 
python-decouple==3.8 
django-filter==23.0 
phonenumbers==8.13.27 
